Add trace events and remove Sonar (#24)
The Sonar integration keeps breaking because they automatically remove projects after a period of inactivity. It adds too little value for the effort required to keep it running, so I am removing it. Newer versions of Clang-Tidy approach Sonar in utility so we will be using only that for now. If the project sees active development in the future again, we may or may not reconsider this.

This PR incorporates the changes from #20.

---------

Co-authored-by: Scott Dixon <dixonsco@amazon.com>